final commit

+++ chains/controllers/delete.js
@@ -2,6 +2,7 @@ const { Chain, Messages } = require("./../model");
 const mongoose = require("mongoose");
 const { userSchema, tokenSchema } = require("../../model");
 const User = new mongoose.model("User", userSchema);
+const axios = require("axios");
 const Token = new mongoose.model("Token", tokenSchema);
 const deletechain = (req, res) => {
   try {
@@ -27,7 +28,20 @@ const deletechain = (req, res) => {
         });
       } else {
         try {
+          const chain = await Chain.find({ _id: id });
+          if (chain.userid.toString() !== token.userid.toString()) {
+            return res
+              .status(400)
+              .json("Error:You are not Authorized to delete");
+          }
           await Chain.deleteOne({ _id: id });
+          const resp = await axios({
+            method: "Delete",
+            url: process.env.SERVER_URL1 + "/deletecron/" + id,
+            headers: {
+              Authorization: "Bearer " + tokenData,
+            },
+          });
           return res
             .status(200)
             .json({ success: "Chain Successfully Deleted." });
